Backport fb46d4e578db297c66176852bec77fc3ad5b89a8
